Use subclasses of container model
