+++ PresentationLayer/UIComponents/Screens/ClaimDevice/ClaimDeviceContainer/ClaimDeviceContainerViewModel.swift
@@ -17,26 +17,22 @@ class ClaimDeviceContainerViewModel: ObservableObject {
 	@Published var loadingState: ClaimDeviceProgressView.State = .loading(.init(title: "", subtitle: ""))
 
 	var steps: [ClaimDeviceStep] = []
-	let navigationTitle: String
+	var navigationTitle: String = ""
 	let useCase: MeUseCase
 	let devicesUseCase: DevicesUseCase
 	let deviceLocationUseCase: DeviceLocationUseCase
+	var claimingKey: String?
+	var serialNumber: String?
+	var location: DeviceLocation?
 
 	private let CLAIMING_RETRIES_MAX = 25 // For 2 minutes timeout
 	private let CLAIMING_RETRIES_DELAY_SECONDS: TimeInterval = 5
-	private var btDevice: BTWXMDevice?
-	private var claimingKey: String?
-	private var serialNumber: String?
-	private var location: DeviceLocation?
-	private var heliumFrequency: Frequency?
 	private var cancellableSet: Set<AnyCancellable> = .init()
 
-	init(type: ClaimStationType, useCase: MeUseCase, devicesUseCase: DevicesUseCase, deviceLocationUseCase: DeviceLocationUseCase) {
-		navigationTitle = type.navigationTitle
+	init(useCase: MeUseCase, devicesUseCase: DevicesUseCase, deviceLocationUseCase: DeviceLocationUseCase) {
 		self.useCase = useCase
 		self.devicesUseCase = devicesUseCase
 		self.deviceLocationUseCase = deviceLocationUseCase
-		steps = getSteps(for: type)
 	}
 
 	func moveNext() {
@@ -54,7 +50,7 @@ class ClaimDeviceContainerViewModel: ObservableObject {
 	}
 }
 
-private extension ClaimDeviceContainerViewModel {
+extension ClaimDeviceContainerViewModel {
 	func moveTo(index: Int) {
 		isMovingNext = index > selectedIndex
 		DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
@@ -62,106 +58,6 @@ private extension ClaimDeviceContainerViewModel {
 		}
 	}
 	
-	func getSteps(for type: ClaimStationType) -> [ClaimDeviceStep] {
-		switch type {
-			case .m5:
-				getM5Steps()
-			case .d1:
-				getD1Steps()
-			case .helium:
-				getHeliumSteps()
-			case .pulse:
-				getPulseSteps()
-		}
-	}
-
-	func getM5Steps() -> [ClaimDeviceStep] {
-		let beginViewModel = ViewModelsFactory.getClaimStationM5BeginViewModel { [weak self] in
-			self?.moveNext()
-		}
-
-		let snViewModel = ViewModelsFactory.getClaimStationM5SNViewModel { [weak self] serialNumber in
-			self?.handleSeriaNumber(serialNumber: serialNumber)
-		}
-
-		let manualSNViewModel = ViewModelsFactory.getManualSNM5ViewModel { [weak self] fields in
-			self?.handleSNInputFields(fields: fields)
-		}
-
-		let locationViewModel = ViewModelsFactory.getClaimDeviceLocationViewModel { [weak self] location in
-			self?.location = location
-			self?.performClaim()
-		}
-
-		return [.begin(beginViewModel), .serialNumber(snViewModel), .manualSerialNumber(manualSNViewModel), .location(locationViewModel)]
-	}
-
-	func getD1Steps() -> [ClaimDeviceStep] {
-		let beginViewModel = ViewModelsFactory.getClaimStationBeginViewModel { [weak self] in
-			self?.moveNext()
-		}
-
-		let snViewModel = ViewModelsFactory.getClaimStationSNViewModel { [weak self] serialNumber in
-			self?.handleSeriaNumber(serialNumber: serialNumber)
-		}
-
-		let manualSNViewModel = ViewModelsFactory.getManualSNViewModel { [weak self] fields in
-			self?.handleSNInputFields(fields: fields)
-		}
-
-		let locationViewModel = ViewModelsFactory.getClaimDeviceLocationViewModel { [weak self] location in
-			self?.location = location
-			self?.performClaim()
-		}
-
-		return [.begin(beginViewModel), .serialNumber(snViewModel), .manualSerialNumber(manualSNViewModel), .location(locationViewModel)]
-	}
-
-	func getHeliumSteps() -> [ClaimDeviceStep] {
-		let resetViewModel = ViewModelsFactory.getResetDeviceViewModel { [weak self] in
-			self?.moveNext()
-		}
-
-		let selectDeviceViewModel = ViewModelsFactory.getSelectDeviceViewModel(useCase: devicesUseCase) { [weak self] (device, error) in
-			guard let self else {
-				return
-			}
-
-			if let error {
-				let object = self.getFailObject(for: error) { [weak self] in
-					self?.showLoading = false
-				}
-				self.loadingState = .fail(object)
-				self.showLoading = true
-			} else {
-				self.btDevice = device
-				self.moveNext()
-			}
-		}
-
-		let setFrequencyViewModel = ViewModelsFactory.getClaimDeviceSetFrequncyViewModel { [weak self] frequency in
-			self?.heliumFrequency = frequency
-			self?.performHeliumClaim()
-		}
-
-		let locationViewModel = ViewModelsFactory.getClaimDeviceLocationViewModel { [weak self, weak setFrequencyViewModel] location in
-			self?.location = location
-			let selectedFrequency = location.getFrequencyFromCurrentLocationCountry(from: self?.deviceLocationUseCase.getCountryInfos())
-			setFrequencyViewModel?.preSelectedFrequency = selectedFrequency
-			setFrequencyViewModel?.selectedFrequency = selectedFrequency
-			setFrequencyViewModel?.selectedLocation = location
-			setFrequencyViewModel?.didSelectFrequencyFromLocation = true
-			self?.moveNext()
-		}
-
-
-		return [.reset(resetViewModel), .selectDevice(selectDeviceViewModel), .location(locationViewModel), .setFrequency(setFrequencyViewModel)]
-	}
-
-	func getPulseSteps() -> [ClaimDeviceStep] {
-		[]
-	}
-
 	func handleSeriaNumber(serialNumber: ClaimDeviceSerialNumberViewModel.SerialNumber?) {
 		guard let serialNumber else {
 			moveNext()
@@ -268,28 +164,6 @@ private extension ClaimDeviceContainerViewModel {
 		return object
 	}
 
-	func getFailObject(for btError: BluetoothHeliumError, retryAction: @escaping VoidCallback) -> FailSuccessStateObject {
-		let uiInfo = btError.uiInfo
-		let object = FailSuccessStateObject(type: .claimDeviceFlow,
-											title: uiInfo.title,
-											subtitle: uiInfo.description?.attributedMarkdown,
-											cancelTitle: LocalizableString.ClaimDevice.cancelClaimButton.localized,
-											retryTitle: LocalizableString.ClaimDevice.retryClaimButton.localized,
-											contactSupportAction: {
-			HelperFunctions().openContactSupport(successFailureEnum: .claimDeviceFlow, email: MainScreenViewModel.shared.userInfo?.email)
-		}) {
-			Router.shared.popToRoot()
-		} retryAction: {
-			WXMAnalytics.shared.trackEvent(.userAction, parameters: [.actionName: .heliumBLEPopupError,
-																	 .contentType: .heliumBLEPopup,
-																	 .action: .tryAgain])
-
-			retryAction()
-		}
-
-		return object
-	}
-
 	func getSuccessObject(for device: DeviceDetails, followState: UserDeviceFollowState?) -> FailSuccessStateObject {
 		let needsUpdate = device.needsUpdate(mainVM: MainScreenViewModel.shared, followState: followState) == true
 		let cancelTitle: String? = needsUpdate ? LocalizableString.ClaimDevice.updateFirmwareAlertGoToStation.localized : nil
@@ -347,98 +221,6 @@ private extension ClaimDeviceContainerViewModel {
 																						  cellCenter: device.cellCenter?.toCLLocationCoordinate2D()))
 			Router.shared.navigateTo(route)
 		}
-	}
-
-	func performHeliumClaim() {
-		guard let btDevice else {
-			return
-		}
-
-		let title = LocalizableString.ClaimDevice.claimingTitle.localized
-		let boldText = LocalizableString.ClaimDevice.claimingTextInformation.localized
-		let subtitle = LocalizableString.ClaimDevice.claimingText(boldText).localized
-
-		Task { @MainActor in
-			var steps: [StepsView.Step] = HeliumSteps.allCases.map { .init(text: $0.description, isCompleted: false) }
-
-			loadingState = .loading(.init(title: title,
-										  subtitle: subtitle.attributedMarkdown,
-										  steps: steps,
-										  stepIndex: 0,
-										  progress: nil))
-			showLoading = true
-
-			// Set frequency
-			if let setFrequencyError = await setHeliumFrequency() {
-				let failObj = self.getFailObject(for: setFrequencyError) { [weak self] in
-					self?.showLoading = false
-				}
-				loadingState = .fail(failObj)
-
-				return
-			}
-
-			steps[0].isCompleted = true
-
-			// Reboot
-
-			loadingState = .loading(.init(title: title,
-										  subtitle: subtitle.attributedMarkdown,
-										  steps: steps,
-										  stepIndex: 1,
-										  progress: nil))
-
-			if let rebootError = await reboot() {
-				let failObj = self.getFailObject(for: rebootError) { [weak self] in
-					self?.showLoading = false
-				}
-
-				loadingState = .fail(failObj)
-
-				return
-			}
-
-			steps[1].isCompleted = true
-
-			// Fetch device info
-			
-			loadingState = .loading(.init(title: title,
-										  subtitle: subtitle.attributedMarkdown,
-										  steps: steps,
-										  stepIndex: 2,
-										  progress: nil))
-
-			let result = await devicesUseCase.getDeviceInfo(device: btDevice)
-			switch result {
-				case .success(let info):
-					// perform claim request
-					self.serialNumber = info?.devEUI
-					self.claimingKey = info?.claimingKey
-					self.performClaim(retries: 0)
-				case .failure(let error):
-					let failObj = self.getFailObject(for: error) { [weak self] in
-						self?.showLoading = false
-					}
-					loadingState = .fail(failObj)
-			}
-		}
-	}
-
-	func setHeliumFrequency() async -> BluetoothHeliumError? {
-		guard let heliumFrequency, let btDevice else {
-			return .unknown
-		}
-
-		let error = await devicesUseCase.setHeliumFrequency(btDevice, frequency: heliumFrequency)
-		return error
-	}
-
-	func reboot() async -> BluetoothHeliumError? {
-		guard let btDevice else {
-			return .unknown
-		}
-
-		return await devicesUseCase.reboot(device: btDevice)
 	}
 }
